% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abracadabra.R
\name{decorate}
\alias{decorate}
\title{Main wrapper}
\usage{
decorate(f)
}
\arguments{
\item{f}{function to be wrapped.}
}
\value{
wrapped function.
}
\description{
Wrapper that automates the seed setting and adjustment of the text in the
main functions of the packages.
}
