% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abracadabra.R
\name{mock_function_for_documentation}
\alias{mock_function_for_documentation}
\title{Mock function for the sole purpose of documenting main functions}
\usage{
mock_function_for_documentation(max.char = 0, flag = 0, seed = 0)
}
\arguments{
\item{max.char}{integer; maximum number of characters to be returned; if NULL (DEFAULT) the text is not cropped.}

\item{flag}{character; if max.char is used, everything after the last time this character appears will be removed; the DEFAULT is a single space " ".}

\item{seed}{numeric; a seed for the draw; the DEFAULT is the system's current day.}
}
\description{
Mock function for the sole purpose of documenting main functions
}
